Changes:
-Redid wrist control implementation (Again). Will probably have to
modify it a bit further.
-Added LUT for Arm angles based on voltage
-DeployIntake will now retract in manual mode until wrist motor current
exceeds a threshold (to be confirmed)
-Made the wrist PIDController depend on the absolute angle, rather than
the relative motor. This will hopefully solve issues wiht a moving
upper/lower limit, but makes it a bit more tricky to use relative
angles.
-LED Code was updated to be a set rather than a toggle. Will potentially
be used to help driver identify cube intake.
-Remapped Operator controls. Cube intake and piston controls have been
moved to the driver, which frees up the co-driver commands to do more
stuff.
-Arm/Elevator actuation has been grouped into a separate command for the
co-driver, to allow for better combination of setpoints.
-Split all toggle piston commands to separate high/low commands to
reduce driver confusion.